make the code browser compatible
It adds babel and browserify to the module's build phase, exposing
a babelified index.js as main entrypoint while retaining src into
new src folder.
