+++ .gitignore
@@ -4,3 +4,4 @@ node_modules
 npm-debug*
 coverage
 .nyc_output
+/dist/*
