+++ package.json
@@ -1,11 +1,15 @@
 {
   "name": "ordinal-js",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "Utility to convert from numbers to their ordinal representations",
   "main": "index.js",
   "author": "Claudio Fernandez <claudio.fernandez@bluealba.com>",
   "license": "ISC",
   "devDependencies": {
+    "babel": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babelify": "^7.3.0",
+    "browserify": "^14.4.0",
     "chai": "^4.1.1",
     "coveralls": "^2.13.1",
     "eslint": "^4.5.0",
@@ -13,8 +17,9 @@
     "nyc": "^11.1.0"
   },
   "scripts": {
-    "test": "eslint '.' && nyc --reporter=html --reporter=text mocha",
-    "lint": "eslint '.'",
+    "build": "babel src --out-dir dist",
+    "test": "eslint './src' && nyc --reporter=html --reporter=text mocha",
+    "lint": "eslint './src'",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "keywords": [
